pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-memory-mcp"
version = "0.1.0"
description = "Unified RAG memory MCP server for ChatGPT + Claude + Grok exports"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastmcp>=0.1.0",
    "qdrant-client>=1.7.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "sparklines>=0.4.0",
    "msgpack>=1.0.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
memory = "cli:main"
memory-indexer = "indexer:main"
memory-server = "server:main"
memory-http = "server_http:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
